Prechat_form/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
python-decouple==3.8
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
